Move restricted unpickler into its own module
Extract RestrictedUnpickler and its lazily-built allowlist out of the
protocol package __init__ into addon/lib/protocol/_restrictedUnpickling.py,
exposed via restrictedLoads(). Keeps the security-sensitive deserialization
logic self-contained and easier to audit; behaviour is unchanged.
